moving helpers for array_and_broadcastable_shape strategy to test_dtype
